+++ CHANGELOG.md
@@ -1,10 +1,9 @@
 # Changelog
 
-## 0.2.1
-  * Integration tests improvements [#31](https://github.com/singer-io/tap-workday/pull/31)
 
 ## 0.2.0
-  * Added OAuth 2.0 authentication support alongside existing username/password auth [#32](https://github.com/singer-io/tap-workday/pull/32)
+  * Added OAuth 2.0 authentication support alongside existing username/password auth
+  * Integration tests improvements [#32](https://github.com/singer-io/tap-workday/pull/32)
 
 ## 0.1.0
   * Supports discovery and sync modes for Workday SOAP API streams